후기 별점, 관리자 통계

+++ semiproject/src/main/java/com/kh/semiproject/mapper/ReviewDetailEditMapper.java
@@ -20,7 +20,7 @@ public class ReviewDetailEditMapper implements RowMapper<ReviewDetailEditDto>{
 				.reviewContent(rs.getString("review_content"))
 				.reviewLike(rs.getInt("review_like"))
 				.reviewRead(rs.getInt("review_read"))
-				.reviewStar(rs.getInt("review_star"))
+				.reviewStar(rs.getFloat("review_star"))
 				.reviewWtime(rs.getTimestamp("review_wtime"))
 				.reviewWriter(rs.getString("review_writer"))
 				.reviewPlace(rs.getInt("review_place"))
